requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
Pillow==10.1.0
daphne==4.0.0

# Authentication and security
django-rest-framework-simplejwt==5.3.0
django-rest-auth==0.9.5
django-allauth==0.58.2

# Additional utilities
python-dotenv==1.0.0
django-environ==0.10.0

# Development and debugging
ipython==8.17.2
django-extensions==3.2.3
werkzeug==3.0.1
